GoogleSignIn, MainMenu, 2DFlashGame, and Settings
GoogleSignIn, should be the initial scene. It utilizes the Google-Sign-In v1.0.4 package.
Main Menu navigates through these scenes. There's one bug that exists, where the settings must be opening first in order for the words to appear in the 2D Flash Game and Flashcards. No storage capabilities yet, just sign in.